refactor: change the name of the page schemas
