+++ frontend/src/app/(auth)/register/register.schema.ts
